# Replace siphash by hash/maphash