--- a/doc/020_installation.rst
+++ b/doc/020_installation.rst
@@ -255,7 +255,7 @@ From Source
 ***********
 
 restic is written in the Go programming language and you need at least
-Go version 1.13. Building restic may also work with older versions of Go,
+Go version 1.14. Building restic may also work with older versions of Go,
 but that's not supported. See the `Getting
 started <https://golang.org/doc/install>`__ guide of the Go project for
 instructions how to install Go.