--- a/go.mod
+++ b/go.mod
@@ -36,4 +36,4 @@ require (
 	gopkg.in/tomb.v2 v2.0.0-20161208151619-d5d1b5820637
 )
 
-go 1.13
+go 1.14