@greatroar 2021-09-18
Oh snap. I thought Go 1.14 was the minimum required version, but it's 1.13. That doesn't have maphash yet.