@MichaelEischer 2021-09-19
Bumping the minimum required go version to 1.14 should be fine. We'll probably soon see libraries requiring 1.14 anyways (as in #3407).