@greatroar 2021-09-19
Added a Go version bump. I resisted the temptation to refactor internal/test to use the new testing.T.Cleanup function :)